move not categorized auth exceptions to cognito auth
