+++ aws-auth-cognito/src/main/java/com/amplifyframework/auth/cognito/exceptions/service/CodeMismatchException.kt
@@ -12,13 +12,13 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package com.amplifyframework.auth.exceptions.service
+package com.amplifyframework.auth.cognito.exceptions.service
 
 import com.amplifyframework.auth.exceptions.ServiceException
 
 /**
- * Could not perform the action because the password given is invalid.
- * @param cause The underlying cause of this exception
+ * Could not perform the action because user entered incorrect confirmation code.
+ * @param cause The underlying cause of the exception
  */
-open class InvalidPasswordException(cause: Throwable?) :
-    ServiceException("The password given is invalid.", "Enter correct password.", cause)
+open class CodeMismatchException(cause: Throwable?) :
+    ServiceException("Confirmation code entered is not correct.", "Enter correct confirmation code.", cause)
